Added example set of git commands
I've added an example set of git commands to edit the specification.  If you have a preferred set of commands, please revise, but these have worked for me.  We cannot assume that editors of the specification are git experts, so it is important to make clear how we want them to proceed.